+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 # 2017-12-20
 
 * Project Started (not really started this date, but this particular Visual Studio Solution was created on this date)
+
+# 2018-01-03
+
+* Added arrayExtensions.removeAt
+
+* Added stringExtensions.copyStringToString and copyStringToCharArray
+
+* Added stringExtensions.textTrim.cs file containing several functions (trimLeft, trimRight, trimBoth, with each function providing overloads that accept either a string-to-trim or a charArray-to-trim)